Allow launching ec2 instances via a %user@instance-type: hostspec
If first char in a hostspec is a % sign (sorry people with user names
starting in %), lunchbox will interpret the hostspec to be a request
for a new EC2 instance.

Currently does not shut down the instance on errors or when
terminating lunchbox.

Requires ec2-api-tools command line tools.

New args --ec2-ami, --ec2-general-args and --ec2-specific-args will be
passed through to ec2-run-instances (--ec2-ami defaults to the alestic
32-bit server EBS AMI).